json: move string-escaping helpers into their own escape module
