+++ src/format/json/mod.rs
@@ -1,2 +1,3 @@
+mod escape;
 pub(crate) mod parser;
 pub(crate) mod writer;
